adding a unit to checkTerraformStateIsEmpty
