+++ apps/wing/src/commands/compile.ts
@@ -10,6 +10,7 @@ import * as wingCompiler from "../wingc";
 import { copyDir, normalPath } from "../util";
 import { CHARS_ASCII, emitDiagnostic, Severity, File, Label } from "codespan-wasm";
 import { existsSync } from "fs";
+import { Target } from "./constants";
 
 // increase the stack trace limit to 50, useful for debugging Rust panics
 // (not setting the limit too high in case of infinite recursion)
@@ -18,18 +19,6 @@ Error.stackTraceLimit = 50;
 const log = debug("wing:compile");
 const WINGC_COMPILE = "wingc_compile";
 const WINGC_PREFLIGHT = "preflight.js";
-
-/**
- * Available targets for compilation.
- * This is passed from Commander to the `compile` function.
- */
-export enum Target {
-  TF_AWS = "tf-aws",
-  TF_AZURE = "tf-azure",
-  TF_GCP = "tf-gcp",
-  SIM = "sim",
-  AWSCDK = "awscdk",
-}
 
 const DEFAULT_SYNTH_DIR_SUFFIX: Record<Target, string | undefined> = {
   [Target.TF_AWS]: "tfaws",
